WIP Separate `Dashboard` class
